refactor(config): move config saving from config controller to config module
