+++ src/icon/config/config.py
@@ -1,12 +1,14 @@
 import logging
+import os
 from pathlib import Path, PosixPath
 
 import yaml
 from confz import BaseConfig, FileSource
 
 from icon.config import latest, v1
-from icon.config.config_path import get_config_path
 from icon.config.migrations import migration_by_version
+
+_ENV_KEY = "ICON_CONFIG"
 
 logger = logging.getLogger("config")
 
@@ -52,3 +54,32 @@ def get_config() -> latest.ServiceConfig:
             file.write(yaml.dump(config.model_dump()))
 
     return config
+
+
+def save_config(config: latest.ServiceConfig) -> None:
+    """Save the configuration to the source YAML file.
+
+    Serializes the updated configuration and writes it back to the file.
+
+    Args:
+        config:
+            The validated configuration instance.
+    """
+    with get_config_path().open("w") as f:
+        f.write(yaml.dump(config.model_dump()))
+
+
+def _normalize(p: str | Path) -> Path:
+    return Path(p).expanduser().resolve()
+
+
+def set_config_path(p: Path) -> None:
+    """Set once at startup; children inherit via environment."""
+    os.environ[_ENV_KEY] = str(_normalize(p))
+
+
+def get_config_path() -> Path:
+    """Read from env, else default."""
+    if env := os.environ.get(_ENV_KEY):
+        return _normalize(env)
+    return _normalize(Path.home() / ".config/icon/config.yaml")
